Create Main.py
Code that was worked at the Fall 2015 Coinorado 24 hour hackathon. This code allows the user to send a text message to a number and receive a text back with the users online bitcoin wallet balance and also real time conversion rates. A server was not set up due to the limiting time constrictions so gmail was used as a mediator.